+++ sponsor/ui/lib/ui/item/cubit/membership_map_item_cubit.dart
@@ -32,6 +32,5 @@ class MembershipMapItemCubit extends Cubit<MembershipMapItemState> {
 
       emit(state);
     });
-    ;
   }
 }