+++ sponsor/ui/lib/ui/setting/cubit/sponsor_setting_item_state.dart
@@ -1,0 +1,12 @@
+
+abstract class SponsorSettingItemState {}
+
+class LoadingSponsorSettingItemState extends SponsorSettingItemState {}
+
+class ErrorSponsorSettingItemState extends SponsorSettingItemState {}
+
+class InfoSponsorSettingItemState extends SponsorSettingItemState {
+  bool isMember;
+
+  InfoSponsorSettingItemState({required this.isMember});
+}